Applied missing changes from thesis.
1. Applied missing changes from thesis (mentioned in project rubric).
2. A minor change not mentioned in the rubric was applied as well;
Modify the current TCB pointer to point to the List Item at the head of xReadyListEDF, i.e. the one with the nearest (smallest) deadline (xItemValue).
Otherwise it would point to the task with the highest priority. If all tasks share the same priority, it would point to the last created task (prior to this update).